+++ Sources/StreamChatUI/Composer/ComposerVC.swift
@@ -13,6 +13,8 @@ public enum AttachmentValidationError: Error {
 
     /// The number of attachments reached the limit.
     case maxAttachmentsCountPerMessageExceeded(limit: Int)
+
+    internal static var fileSizeMaxLimitFallback: Int64 = 100 * 1024 * 1024
 }
 
 public struct LocalAttachmentInfoKey: Hashable, Equatable, RawRepresentable {
@@ -1201,7 +1203,8 @@ open class ComposerVC: _ViewController,
         }
 
         let fileSize = try AttachmentFile(url: url).size
-        guard fileSize < chatConfig.maxAttachmentSize else {
+        let maxAttachmentSize = maxAttachmentSize(for: type)
+        guard fileSize <= maxAttachmentSize else {
             throw AttachmentValidationError.maxFileSizeExceeded
         }
 
@@ -1229,6 +1232,35 @@ open class ComposerVC: _ViewController,
             extraData: extraData
         )
         content.attachments.append(attachment)
+    }
+
+    /// The maximum upload file size depending on the attachment type.
+    ///
+    /// The max attachment size can be set from the Stream's Dashboard App Settings.
+    /// If it is not set, it fallbacks to the deprecated `ChatClientConfig.maxAttachmentSize`.
+    /// - Parameter attachmentType: The attachment type that is being uploaded.
+    /// - Returns: The file size limit in bytes. The default value is 100MB.
+    open func maxAttachmentSize(for attachmentType: AttachmentType) -> Int64 {
+        guard let client = channelController?.client else {
+            log.assertionFailure("Channel controller must be set at this point")
+            return AttachmentValidationError.fileSizeMaxLimitFallback
+        }
+
+        let maxAttachmentSize: Int64?
+        switch attachmentType {
+        case .image:
+            maxAttachmentSize = client.appSettings?.imageUploadConfig.sizeLimitInBytes
+        default:
+            maxAttachmentSize = client.appSettings?.fileUploadConfig.sizeLimitInBytes
+        }
+
+        // If no value is set in the dashboard, the size_limit will be nil or zero,
+        // so in this case we fallback to the deprecated value.
+        guard let maxSize = maxAttachmentSize, maxSize > 0 else {
+            return client.config.maxAttachmentSize
+        }
+
+        return maxSize
     }
 
     /// Shows an alert for the error thrown when adding attachment to a composer.